refactor: DispatcherServletInitializer App 패키지로 이동
