+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,9 +1,10 @@
-package webmvc.org.springframework.web.servlet;
+package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.DispatcherServlet;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.AnnotationHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.handler.AnnotationHandlerMapper;
